Implement opus decoder

+++ src/scripts/packets.ts
@@ -41,7 +41,7 @@ export type PlayerStatePacket = PacketPlayerState // voicechat:player_state
 export type SoundPacket = {
     channelId: UUID;
     sender: UUID;
-    data: Float32Array;
+    data: Uint8Array;
     sequenceNumber: Long;
     category?: string;
 }
@@ -56,7 +56,7 @@ export type UpdateStatePacket = {
 }
 
 export type MicPacket = { // 0x01
-    data: Float32Array;
+    data: Uint8Array;
     whispering: boolean;
     sequenceNumber: Long;
 }
